requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
scikit-learn>=1.3
xgboost>=2.0
matplotlib>=3.7
openpyxl>=3.1